requirements.txt
youtube-transcript-api>=0.6.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
python-dotenv>=1.0.0
openai>=1.3.0
streamlit>=1.22.0
